Make waitUntilElementIsAvailable globally available
